Make Transliteration properly load into the client (browser)
The minimized library had obfusicated code, and therefore the export was
no longer Transliteration.  It was "l".  So the meteor
`lib.exports("Transliteration")` wasn't working at all.  The module WAS
forcible putting itself into window.TR and window.Transliteration
but then Meteor blew away the Transliteration immediately with th export.

This changes this to export properly in Meteor fashion.  The file will
be minimized by Meteor.  Only a couple slight changes to the actual
source file.